Implement Driver API for dynamic iPXE config
This patch adds two new abstract mixin classes, DeployIPXEMixin and
BootIPXEMixin, which define methods to fetch options/configs needed for
dynamic iPXE configuration from deploy and boot interfaces respectively.

Implementations are added to AgentDeployMixin and PXEBoot interfaces,
so that both Agent and ISCSI deploy can be used with dynamic
iPXE configuration.

New config option `[pxe]ipxe_server_enabled` is added which enables
some discrimination for when do or do not create pxe configs on disk and
which PXE DHCP options to set.
Later it will be used on ironic-api side as well.

This option is set to deprecated right away, as it sole purpose is to
give operatos that use custom out-of-tree iPXE boot scripts and configs
time to update those to proper Jinja templates.
After deprecation period it should be merged into `[pxe]ipxe_enabled`
option to use dynamic iPXE configuration by default in iPXE environments.

The sample of ironic.conf file is not updated on purpose to not confuse
users, as this option is not yet to be enabled until API-side changes
are implemented.

Change-Id: I259f356fc046f89e28f7ca4ef7301113b3c752ef
Partial-Bug: #1526275